Detect full waitlists correctly, and present an appropriate error message.
A Minerva error message beginning with "Open" is interpreted as an
indication to register in the waitlist section.
"Open - Waitlist full" could be detected at that stage, but we might as
well try to register if Minerva will allow it.
Minerva most likely won't -- we won't find a field we can set to LW.
We can then let the user know that we couldn't get into the course.